tweak some fixture hunk names.
See commit message in the linked commit in the ipld/ipld submodule.

(Calling both the dagjose and the dagjson hunks "data", as it was
previously, suggested incorrect things: they are not that similar.
The dagjson is the datamodel view, and that is *not* the same thing
as a transliteration of the tokens that you'd get if you parsed the
dag-jose as dag-cbor.)